requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
selenium==4.4.2
